Class for 6/29
Introduction to Feign Client, setup Swagger UI,
JWT and implementing config using OpenFeign Client to call other microservices to get cumulative returns for a given stock in a given timeframe.
Implemented an API to get the Top N performing stocks in the market for a given timeframe

+++ src/main/java/io/endeavour/stocks/vo/StockFundamentalsHistoryVO.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class StockFundamentalsHistoryVO
 {
     private String tickerSymbol;
-    private double currentRatio;
+    private BigDecimal currentRatio;
     private BigDecimal marketCap;
     private List<StockPriceHistoryVo> tradingHistory;
 
@@ -19,11 +19,11 @@ public class StockFundamentalsHistoryVO
         this.tickerSymbol = tickerSymbol;
     }
 
-    public double getCurrentRatio() {
+    public BigDecimal getCurrentRatio() {
         return currentRatio;
     }
 
-    public void setCurrentRatio(double currentRatio) {
+    public void setCurrentRatio(BigDecimal currentRatio) {
         this.currentRatio = currentRatio;
     }
